+++ rust/Cargo.toml
@@ -3,7 +3,9 @@ name = "dylibso-observe-sdk"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+# not testing docs because the prost generated structs fail
+doctest = false
 
 [dependencies]
 anyhow = "1"
